More adjustments to #include changes. Moved endian, hash functions out of util.h
