+++ tiledb/sm/misc/utils.cc
@@ -158,135 +158,6 @@ Status get_fragment_version(const std::string& name, uint32_t* version) {
   return Status::Ok();
 }
 
-bool is_int(const std::string& str) {
-  // Check if empty
-  if (str.empty())
-    return false;
-
-  // Check first character
-  if (str[0] != '+' && str[0] != '-' && !(bool)isdigit(str[0]))
-    return false;
-
-  // Check rest of characters
-  for (size_t i = 1; i < str.size(); ++i)
-    if (!(bool)isdigit(str[i]))
-      return false;
-
-  return true;
-}
-
-bool is_uint(const std::string& str) {
-  // Check if empty
-  if (str.empty())
-    return false;
-
-  // Check first character
-  if (str[0] != '+' && !isdigit(str[0]))
-    return false;
-
-  // Check characters
-  for (size_t i = 1; i < str.size(); ++i)
-    if (!(bool)isdigit(str[i]))
-      return false;
-
-  return true;
-}
-
-template <class T>
-std::string to_str(const T& value) {
-  std::stringstream ss;
-  ss << value;
-  return ss.str();
-}
-
-std::string to_str(const void* value, Datatype type) {
-  std::stringstream ss;
-  switch (type) {
-    case Datatype::INT8:
-      ss << *(const int8_t*)value;
-      break;
-    case Datatype::UINT8:
-      ss << *(const uint8_t*)value;
-      break;
-    case Datatype::INT16:
-      ss << *(const int16_t*)value;
-      break;
-    case Datatype::UINT16:
-      ss << *(const uint16_t*)value;
-      break;
-    case Datatype::INT32:
-      ss << *(const int32_t*)value;
-      break;
-    case Datatype::UINT32:
-      ss << *(const uint32_t*)value;
-      break;
-    case Datatype::INT64:
-      ss << *(const int64_t*)value;
-      break;
-    case Datatype::UINT64:
-      ss << *(const uint64_t*)value;
-      break;
-    case Datatype::FLOAT32:
-      ss << *(const float*)value;
-      break;
-    case Datatype::FLOAT64:
-      ss << *(const double*)value;
-      break;
-    case Datatype::CHAR:
-      ss << *(const char*)value;
-      break;
-    case Datatype::ANY:
-      ss << *(const uint8_t*)value;
-      break;
-    case Datatype::STRING_ASCII:
-      ss << *(const uint8_t*)value;
-      break;
-    case Datatype::STRING_UTF8:
-      ss << *(const uint8_t*)value;
-      break;
-    case Datatype::STRING_UTF16:
-      ss << *(const uint16_t*)value;
-      break;
-    case Datatype::STRING_UTF32:
-      ss << *(const uint32_t*)value;
-      break;
-    case Datatype::STRING_UCS2:
-      ss << *(const uint16_t*)value;
-      break;
-    case Datatype::STRING_UCS4:
-      ss << *(const uint32_t*)value;
-      break;
-    case Datatype::DATETIME_YEAR:
-    case Datatype::DATETIME_MONTH:
-    case Datatype::DATETIME_WEEK:
-    case Datatype::DATETIME_DAY:
-    case Datatype::DATETIME_HR:
-    case Datatype::DATETIME_MIN:
-    case Datatype::DATETIME_SEC:
-    case Datatype::DATETIME_MS:
-    case Datatype::DATETIME_US:
-    case Datatype::DATETIME_NS:
-    case Datatype::DATETIME_PS:
-    case Datatype::DATETIME_FS:
-    case Datatype::DATETIME_AS:
-    case Datatype::TIME_HR:
-    case Datatype::TIME_MIN:
-    case Datatype::TIME_SEC:
-    case Datatype::TIME_MS:
-    case Datatype::TIME_US:
-    case Datatype::TIME_NS:
-    case Datatype::TIME_PS:
-    case Datatype::TIME_FS:
-    case Datatype::TIME_AS:
-      ss << *(const int64_t*)value;
-      break;
-    default:
-      assert(false);
-  }
-
-  return ss.str();
-}
-
 uint64_t common_prefix_size(
     const std::basic_string_view<char>& a,
     const std::basic_string_view<char>& b) {
