+++ tiledb/sm/rest/rest_client.cc
@@ -51,7 +51,8 @@
 #include "tiledb/sm/array/array.h"
 #include "tiledb/sm/enums/query_type.h"
 #include "tiledb/sm/misc/constants.h"
-#include "tiledb/sm/misc/utils.h"
+#include "tiledb/sm/misc/endian.h"
+#include "tiledb/sm/misc/parse_argument.h"
 #include "tiledb/sm/query/query.h"
 #include "tiledb/sm/rest/rest_client.h"
 #include "tiledb/sm/serialization/array_schema.h"
